git-svn-id: https://svn.code.sf.net/p/kaldi/code/sandbox/tacc_kaldi@3411 5e6a8d80-dfce-4ca6-a32a-6e07a63d50c8

+++ egs/babel/s5b/run-4-test.sh
@@ -13,6 +13,7 @@ data_only=false
 fast_path=true
 skip_kws=false
 skip_stt=false
+skip_scoring=false
 max_states=150000
 wip=0.5
 . utils/parse_options.sh
@@ -40,19 +41,13 @@ function make_plp {
   if [ "$use_pitch" = "false" ] && [ "$use_ffv" = "false" ]; then
    steps/make_plp.sh --cmd "$decode_cmd" --nj $my_nj data/${t} exp/make_plp/${t} plp
   elif [ "$use_pitch" = "true" ] && [ "$use_ffv" = "true" ]; then
-    cp -rT data/${t} data/${t}_plp; cp -rT data/${t} data/${t}_pitch; cp -rT data/${t} data/${t}_ffv
-    steps/make_plp.sh --cmd "$decode_cmd" --nj $my_nj data/${t}_plp exp/make_plp/${t} plp_tmp_${t}
-    steps/make_pitch_kaldi.sh --cmd "$decode_cmd" --nj $my_nj data/${t}_pitch exp/make_pitch/${t} pitch_tmp_${t}
+    cp -rT data/${t} data/${t}_plp_pitch; cp -rT data/${t} data/${t}_ffv
+    steps/make_plp_pitch.sh --cmd "$decode_cmd" --nj $my_nj data/${t}_plp_pitch exp/make_plp_pitch/${t} plp_pitch_tmp_${t}
     local/make_ffv.sh --cmd "$decode_cmd"  --nj $my_nj data/${t}_ffv exp/make_ffv/${t} ffv_tmp_${t}
-    steps/append_feats.sh --cmd "$decode_cmd" --nj $my_nj data/${t}{_plp,_pitch,_plp_pitch} exp/make_pitch/append_${t}_pitch plp_tmp_${t}
     steps/append_feats.sh --cmd "$decode_cmd" --nj $my_nj data/${t}{_plp_pitch,_ffv,} exp/make_ffv/append_${t}_pitch_ffv plp
-    rm -rf {plp,pitch,ffv}_tmp_${t} data/${t}_{plp,pitch,plp_pitch}
+    rm -rf {plp_pitch,ffv}_tmp_${t} data/${t}_{plp_pitch,ffv}
   elif [ "$use_pitch" = "true" ]; then
-    cp -rT data/${t} data/${t}_plp; cp -rT data/${t} data/${t}_pitch
-    steps/make_plp.sh --cmd "$decode_cmd" --nj $my_nj data/${t}_plp exp/make_plp/${t} plp_tmp_${t}
-    steps/make_pitch_kaldi.sh --cmd "$decode_cmd" --nj $my_nj data/${t}_pitch exp/make_pitch/${t} pitch_tmp_${t}
-    steps/append_feats.sh --cmd "$decode_cmd" --nj $my_nj data/${t}{_plp,_pitch,} exp/make_pitch/append_${t} plp
-    rm -rf {plp,pitch}_tmp_${t} data/${t}_{plp,pitch}
+    steps/make_plp_pitch.sh --cmd "$decode_cmd" --nj $my_nj data/${t} exp/make_plp_pitch/${t} plp
   elif [ "$use_ffv" = "true" ]; then
     cp -rT data/${t} data/${t}_plp; cp -rT data/${t} data/${t}_ffv
     steps/make_plp.sh --cmd "$decode_cmd" --nj $my_nj data/${t}_plp exp/make_plp/${t} plp_tmp_${t}
@@ -64,7 +59,8 @@ function make_plp {
   utils/fix_data_dir.sh data/${t}
 }
 
-if [ ${type} == shadow ] ; then
+#if [ ${type} == shadow ] || [ ${type} == eval ] ; then
+if  [[ ${type} == shadow || $type == eval || $type == semitrain || $type == devtrain ]] ; then
   mandatory_variables=""
   optional_variables=""
 else
@@ -85,6 +81,18 @@ eval my_kwlist_file=\$${type}_kwlist_file
 eval my_rttm_file=\$${type}_rttm_file
 eval my_nj=\$${type}_nj  #for shadow, this will be re-set when appropriate
 
+declare -A my_more_kwlists
+eval my_more_kwlist_keys="\${!${type}_more_kwlists[@]}"
+declare -p my_more_kwlist_keys
+for key in $my_more_kwlist_keys  # make sure you include the quotes there
+do
+  echo $key
+  eval my_more_kwlist_val="\${${type}_more_kwlists[$key]}"
+  echo $my_more_kwlist_val
+  my_more_kwlists["$key"]="${my_more_kwlist_val}"
+done
+
+
 for variable in $mandatory_variables ; do
   eval my_variable=\$${variable}
   if [ -z $my_variable ] ; then
@@ -103,7 +111,7 @@ done
 
 datadir=data/${type}.uem
 dirid=${type}.uem
-
+skip_scoring=false
 if [[ $type == shadow ]] ; then
   if [ ! -f ${datadir}/.done ]; then
     # we expect that the ${dev2shadow} as well as ${eval2shadow} already exist
@@ -123,13 +131,53 @@ if [[ $type == shadow ]] ; then
     touch ${datadir}/.done
   fi
   my_nj=$eval_nj
-else
+elif [[ $type == eval || $type == semitrain || $type == devtrain ]] ; then
   if [ ! -d data/raw_${type}_data ]; then
     echo ---------------------------------------------------------------------
     echo "Subsetting the ${type} set"
     echo ---------------------------------------------------------------------
     
+    local/make_corpus_subset.sh --ignore-missing-txt true "$my_data_dir" "$my_data_list" ./data/raw_${type}_data
+  fi
+  my_data_dir=`readlink -f ./data/raw_${type}_data`
+
+  skip_scoring=true
+
+  nj_max=`cat $my_data_list | wc -l`
+  if [[ "$nj_max" -lt "$my_nj" ]] ; then
+    echo "The maximum reasonable number of jobs is $nj_max -- you have $my_nj! (The training and decoding process has file-granularity)"
+    exit 1
+    my_nj=$nj_max
+  fi
+  
+  if [ ! -f ${datadir}/.done ]; then
+    if [[ ! -f ${datadir}/wav.scp || ${datadir}/wav.scp -ot "$my_data_cmudb" ]]; then
+      echo ---------------------------------------------------------------------
+      echo "Preparing ${type} data lists in ${datadir} on" `date`
+      echo ---------------------------------------------------------------------
+      mkdir -p ${datadir}
+      local/cmu_uem2kaldi_dir.sh --filelist $my_data_list \
+        $my_data_cmudb  $my_data_dir ${datadir}
+    fi
+
+    if [ ! -f ${datadir}/.plp.done ]; then
+      echo ---------------------------------------------------------------------
+      echo "Preparing ${type} parametrization files in ${datadir} on" `date`
+      echo ---------------------------------------------------------------------
+      make_plp ${dirid}
+      touch ${datadir}/.plp.done
+    fi
+
+    touch ${datadir}/.done
+  fi
+else
+  if [ ! -f data/raw_${type}_data/.done ]; then
+    echo ---------------------------------------------------------------------
+    echo "Subsetting the ${type} set"
+    echo ---------------------------------------------------------------------
+    
     local/make_corpus_subset.sh "$my_data_dir" "$my_data_list" ./data/raw_${type}_data
+    touch data/raw_${type}_data/.done
   fi
   my_data_dir=`readlink -f ./data/raw_${type}_data`
 
@@ -199,8 +247,9 @@ if ! $skip_kws  && [ ! -f ${datadir}/kws/.done ] ; then
 
     local/kws_data_prep.sh --case_insensitive $case_insensitive \
       "${icu_opt[@]}" \
-      data/lang ${datadir} ${datadir}/kws
+      data/lang ${datadir} ${datadir}/kws || exit 1
     utils/fix_data_dir.sh ${datadir}
+
 
     touch ${datadir}/kws/.done
   else
@@ -214,10 +263,31 @@ if ! $skip_kws  && [ ! -f ${datadir}/kws/.done ] ; then
     
     local/kws_setup.sh --case_insensitive $case_insensitive \
       "${kws_flags[@]}" "${icu_opt[@]}" \
-      $my_ecf_file $my_kwlist_file data/lang ${datadir}
+      $my_ecf_file $my_kwlist_file data/lang ${datadir} || exit 1
 
+    if [ ${#my_more_kwlists[@]} -ne 0  ] ; then
+      touch $datadir/extra_kws_tasks
+      for extraid in "${!my_more_kwlists[@]}" ; do
+        kwlist=${my_more_kwlists[$extraid]}
+        local/kws_setup.sh --rttm-file ${datadir}/kws/rttm --extraid $extraid \
+          --case_insensitive $case_insensitive "${icu_opt[@]}"  \
+          ${datadir}/kws/ecf.xml $kwlist data/lang ${datadir} || exit 1
+        echo $extraid >> $datadir/extra_kws_tasks;  sort -u $datadir/extra_kws_tasks -o  $datadir/extra_kws_tasks
+      done
+    fi
     touch ${datadir}/kws/.done
   fi
+
+  langid=`ls -1 data/raw_${type}_data/audio/ | head -n 1| cut -d '_' -f 3`
+  local/kws_setup.sh --kwlist-wordlist true --rttm-file $datadir/kws/rttm  --extraid fullvocab  \
+    $datadir/kws/ecf.xml <(cat data/lang/words.txt | grep -v -F "<" | grep -v -F "#"  | \
+                            awk "{printf \"KWID$langid-FULLVOCAB-%05d %s\\n\", \$2, \$1 }" ) \
+    data/lang ${datadir} || exit 1
+  echo fullvocab >> $datadir/extra_kws_tasks;  sort -u $datadir/extra_kws_tasks -o  $datadir/extra_kws_tasks
+
+  #local/make_lexicon_subset.sh $data_dir/transcription $lexicon_file $datadir/filtered_lexicon.txt
+  #local/kws_setup.sh --kwlist-wordlist true --rttm-file $datadir/kws/rttm  --extraid vocab  \
+  #  $datadir/kws/ecf.xml  <(cut -f 1 $datadir/filtered_lexicon.txt  | grep -v -F "<" | grep -v -F "#"  | awk "{printf \"KWID$langid-VOCAB-%05d data/raw_${type}_data/.done %s\\n\", \$2, \$1 }" ) data/lang ${datadir}
 fi
 
 if $data_only ; then
@@ -249,12 +319,12 @@ if [ ! -f ${decode}/.done ]; then
 fi
 
 if ! $fast_path ; then
-  local/run_kws_stt_task.sh --cer $cer --max-states $max_states \
+  local/run_kws_stt_task.sh --cer $cer --max-states $max_states --skip-scoring $skip_scoring\
     --cmd "$decode_cmd" --skip-kws $skip_kws --skip-stt $skip_stt --wip $wip \
     "${shadow_set_extra_opts[@]}" "${lmwt_plp_extra_opts[@]}" \
     ${datadir} data/lang ${decode}
 
-  local/run_kws_stt_task.sh --cer $cer --max-states $max_states \
+  local/run_kws_stt_task.sh --cer $cer --max-states $max_states --skip-scoring $skip_scoring\
     --cmd "$decode_cmd" --skip-kws $skip_kws --skip-stt $skip_stt --wip $wip \
     "${shadow_set_extra_opts[@]}" "${lmwt_plp_extra_opts[@]}" \
     ${datadir} data/lang ${decode}.si
@@ -279,7 +349,7 @@ if [ ! -f $decode/.done ]; then
 fi
 
 if ! $fast_path ; then
-  local/run_kws_stt_task.sh --cer $cer --max-states $max_states \
+  local/run_kws_stt_task.sh --cer $cer --max-states $max_states --skip-scoring $skip_scoring\
     --cmd "$decode_cmd" --skip-kws $skip_kws --skip-stt $skip_stt --wip $wip \
     "${shadow_set_extra_opts[@]}" "${lmwt_plp_extra_opts[@]}" \
     ${datadir} data/lang  exp/sgmm5/decode_fmllr_${dirid}
@@ -311,7 +381,7 @@ done
 for iter in 1 2 3 4; do
   # Decode SGMM+MMI (via rescoring).
   decode=exp/sgmm5_mmi_b0.1/decode_fmllr_${dirid}_it$iter
-  local/run_kws_stt_task.sh --cer $cer --max-states $max_states \
+  local/run_kws_stt_task.sh --cer $cer --max-states $max_states --skip-scoring $skip_scoring\
     --cmd "$decode_cmd" --skip-kws $skip_kws --skip-stt $skip_stt --wip $wip \
     "${shadow_set_extra_opts[@]}" "${lmwt_plp_extra_opts[@]}" \
     ${datadir} data/lang $decode
@@ -326,18 +396,46 @@ if [ -f exp/tri6_nnet/.done ]; then
   decode=exp/tri6_nnet/decode_${dirid}
   if [ ! -f $decode/.done ]; then
     mkdir -p $decode
-    steps/decode_nnet_cpu.sh --cmd "$decode_cmd" --nj $my_nj \
+    steps/nnet2/decode.sh --cmd "$decode_cmd" --nj $my_nj \
+      --beam $dnn_beam --lat-beam $dnn_lat_beam \
       --skip-scoring true "${decode_extra_opts[@]}" \
       --transform-dir exp/tri5/decode_${dirid} \
-      exp/tri5/graph ${datadir} $decode |tee $decode/decode.log
+      exp/tri5/graph ${datadir} $decode | tee $decode/decode.log
 
     touch $decode/.done
   fi
 
-  local/run_kws_stt_task.sh --cer $cer --max-states $max_states \
+  local/run_kws_stt_task.sh --cer $cer --max-states $max_states --skip-scoring $skip_scoring\
     --cmd "$decode_cmd" --skip-kws $skip_kws --skip-stt $skip_stt --wip $wip \
     "${shadow_set_extra_opts[@]}" "${lmwt_plp_extra_opts[@]}" \
     ${datadir} data/lang $decode
+fi
+
+####################################################################
+##
+## DNN_MPE decoding
+##
+####################################################################
+if [ -f exp/tri6_nnet_mpe/.done ]; then
+  for epoch in 1 2 3 4; do
+    decode=exp/tri6_nnet_mpe/decode_${dirid}_epoch$epoch
+    if [ ! -f $decode/.done ]; then
+      mkdir -p $decode
+      steps/nnet2/decode.sh \
+        --cmd "$decode_cmd" --nj $my_nj --iter epoch$epoch \
+        --beam $dnn_beam --lat-beam $dnn_lat_beam \
+        --skip-scoring true "${decode_extra_opts[@]}" \
+        --transform-dir exp/tri5/decode_${dirid} \
+        exp/tri5/graph ${datadir} $decode | tee $decode/decode.log
+
+      touch $decode/.done
+    fi
+
+    local/run_kws_stt_task.sh --cer $cer --max-states $max_states --skip-scoring $skip_scoring\
+      --cmd "$decode_cmd" --skip-kws $skip_kws --skip-stt $skip_stt --wip $wip \
+      "${shadow_set_extra_opts[@]}" "${lmwt_dnn_extra_opts[@]}" \
+      ${datadir} data/lang $decode
+  done
 fi
 
 echo "Everything looking good...."
